// App/enums/FormConstants.ts
export enum FormConstants {
  Email = 'email',
  NewEmail = 'newemail',
  Email2 = 'email2',
  BusinessEmail = 'BusinessEmail',
  Password = 'password',
  OTP = 'otp',
  FirstName = 'first_name',
  LastName = 'last_name',
  Passport = 'passport',
  MiddleName = 'middle_name',
  Gender = 'gender',
  Country = 'country',
  CountryUBO = 'CountryUBO',
  Comment = 'Comment',
  DocumentType = 'doc_type_id',
  ProofId = 'doc_unique_no',
  DOB = 'dob',
  IssueDate = 'issueDate',
  ExpiryDate = 'expiryDate',
  Tax = 'tax_no',
  Name = 'name',
  AccountType = 'account_type',
  KYCStatus = 'kyc_status',
  KYBStatus = 'kyb_status',
  PORStatus = 'por_status',
  RefCode = 'referral_code',
  Pair = 'pair',
  Side = 'side',
  Volume = 'volume',
  Balance = 'balance',
  Total = 'total',
  EstimatedPrice = 'estimated_price',
  EstimatedTotal = 'estimated_total',
  EstimatedFees = 'estimated_fees',
  StreetAddress = 'street_address',
  StreetAddress2 = 'street_address2',
  City = 'city',
  PostalCode = 'postal_code',
  State = 'state',
  Parent1 = 'parent1',
  Parent2 = 'parent2',
  CountryCode = 'mobile_isd',
  PhoneNumber = 'mobile_no',
  Phone1 = 'mobile_no_1',
  Phone2 = 'mobile_no_2',
  CurrentPassword = 'oldPassword',
  NewPassword = 'newPassword',
  ConfirmNewPassword = 'confirmPassword',
  Asset = 'asset',
  AssetPriceInUsd = 'assetPrice',
  BonusPercentage = 'bonusPercentage',
  BonusTotal = 'bonusTotal',
  AssetPriceSecondary = 'assetPriceSecondary',
  SecondaryCurrency = 'secondaryCurrency',
  Address = 'address',
  Amount = 'amount',
  BankName = 'bank',
  BankAddress = 'bank_address',
  BankCode = 'beneficiaryBankCode',
  BankBranchCode = 'beneficiaryBankBranch',
  SwiftCode = 'swift_bic_code',
  AccountNumber = 'ibn_account',
  Fee = 'fee',
  FinalAmount = 'final_amount',
  Currency = 'currency',
  ClientId = 'clientId',
  AvailableCurrency = 'availableCurrency',
  Type = 'type',
  Beneficiary = 'beneficiary',
  CurrencyFee = 'currencyFee',
  Currency_Label = 'currency_label',
  FeeAmount = 'fee_amount',
  BeneficiaryName = 'beneficiaryName',
  BeneficiaryAddress = 'address',
  BeneficiaryBankName = 'bank_name',
  TFA = 'tfa',
  CompanyName = 'companyName',
  RegisterationNo = 'companyNo',
  RegisterationDate = 'regDate',
  Vat = 'vat_number',
  TaxNo = 'tax_number',
  DestinationTag = 'destTag',
  TFASetupEmailOTP = 'tfa_setup_email_otp',
  TFASetupCode = 'tfa_setup_code',
  TFAStatus = 'tfaStatus',
  Reference = 'reference',
  StandardCurrency = 'standard_currency',
  IsVat = 'isVat',
  VatPercent = 'vatPercent',
  VatText = 'vatText',
  MerchantAddress = 'merchantAddress',
  MerchantDesc = 'merchantDesc',
  Industry = 'industry',
  ThisDay = 'thisDay',
  ThisMonth = 'thisMonth',
  ThisYear = 'thisYear',
  Tips = 'tip',
  FromDate = 'from_date',
  EndDate = 'end_date',
  IndividualBusinessROle = 'individual_business_role',
  BusinessType = 'type_of_business',
  BusinessName = 'business_name',
  Licensed = 'licensed',
  LicenseType = 'license_type',
  LicenseNumber = 'license_number',
  LicenseComments = 'license_comments',
  RoleOfOwner = 'role_of_owner',
  companyCert = 'companyCert',
  memoAndArticle = 'memoAndArticle',
  registerOfShareholders = 'registerOfShareholders',
  registerOfOfficer = 'registerOfOfficer',
  bankStatement = 'bankStatement',
  meetingToOpen = 'meetingToOpen',
  Stake = 'Stake',
  Wallet = 'Wallet',
  OrderAS = 'OrderAS',
  From = 'From',
  Channel = 'Channel',
  CardType = 'CardType',
  Title = 'Title',
  Nationality = 'Nationality',
  MaritalStatus = 'MaritalStatus',
  EmergencyContactName = 'EmergencyContactName',
  BiometricImage = 'BiometricImage',
  PassportImage = 'PassportImage',
  SignatureImage = 'SignatureImage',
  PictureSelfieImage = 'PictureSelfieImage',
  MailingAddress = 'MailingAddress',
  MailingCity = 'MailingCity',
  MailingState = 'MailingState',
  MailingPostalCode = 'MailingPostalCode',
  MailingCountry = 'MailingCountry',
  Delivery = 'Delivery',
  FirstDeposit = 'FirstDeposit',
  Card = 'Card',
  CardNo = 'CardNo',
  PIN = 'PIN',
  CVV = 'CVV',
  Expiry = 'Expiry',
  Voucher = 'Voucher',
  Region = 'Region',
  BSB = 'BSB',
  PayIDName = 'PayIDName',
  PayID = 'PayID',
  Status = 'Status',
}