// App/enums/Screen.ts
export enum Screen {
  Home = 'Home',
  App = 'App',
  Login = 'Login',
  SelectRegion = 'SelectRegion',
  Auth = 'Auth',
  Tickers = 'Tickers',
  Derivatives = 'Derivatives',
  Wallets = 'Wallets',
  Account = 'Account',
  Funding = 'Funding',
  KycProcess = 'KYCProcess',
  VerifyKyc = 'VerifyKyc',
  MyAccount = 'MyAccount',
  QuickBuy = 'QuickBuy',
  QuickBuySpot = 'QuickBuySpot',
  QuickBuyScreen = 'QuickBuyScreen',
  WalletScreen = 'WalletScreen',
  TickerScreen = 'TickerScreen',
  EditProfile = 'EditProfile',
  PORScreen = 'PORScreen',
  BusinessScreen = 'BusinessScreen',
  IdentityVerificationScreen = 'IdentityVerificationScreen',
  ChangePassword = 'ChangePassword',
  Signup = 'Signup',
  Statistics = 'Statistics',
  WebViewWrapper = 'WebViewWrapper',
  ChooseLanguage = 'ChooseLanguage',
  SelectLanguage = 'SelectLanguage',
  Security = 'Security',
  SetupPasscode = 'SetupPasscode',
  BioAuthNavigator = 'BioAuthNavigator',
  ConfigureBioAuth = 'ConfigureBioAuth',
  VerifyBioAuth = 'VerifyBioAuth',
  VerifyPasscode = 'VerifyPasscode',
  AuthLoading = 'AuthLoading',
  CurrentPasscode = 'CurrentPasscode',
  ChangePasscode = 'ChangePasscode',
  Theme = 'Theme',
  VerifyModal = 'VerifyModal',
  GbexScreen = 'GbexScreen',
  Gbex = 'Gbex',
  Coingate = 'Coingate',
  CardMercuryo = 'CardMercuryo',
  CryptoPayment = 'CryptoPayment',
  Payment = 'Payment',
  Beneficiary = 'Beneficiary',
  IBANS = 'IBANS',
  PayID = 'PayID',
  InternalPayment = 'InternalPayment',
  Deposit = 'Deposit',
  Withdrawal = 'Withdrawal',
  CreditCardDeposit = 'CreditCardDeposit',
  GbexHistory = 'GbexHistory',
  VerifyKyb = 'VerifyKyb',
  MainCurrency = 'MainCurrency',
  SecondCurrency = 'SecondCurrency',
  HomeTab = 'HomeTab',
  GbexStatusScreen = 'GbexStatusScreen',
  Rewards = 'Rewards',
  Dashboard = 'Dashboard',
  RewardHistory = 'RewardHistory',
  Limit = 'Limit',
  ForgotPasswordEnterEmail = 'ForgotPasswordEnterEmail',
  ForgotPasswordEnterPassword = 'ForgotPasswordEnterPassword',
  TwoFactorAuthentication = 'TwoFactorAuthentication',
  GlobiancePos = 'GlobiancePos',
  QrPay = 'QrPay',
  QrPayScan = 'QrPayScan',
  Checkout = 'Checkout',
  CheckoutHome = 'CheckoutHome',
  PosConfig = 'PosConfig',
  Merchant = 'Merchant',
  PosHistoryDetail = 'PosHistoryDetail',
  PosHistoryExport = 'PosHistoryExport',
  QuickSwap = 'QuickSwap',
  QuickSwapScreen = 'QuickSwapScreen',
  CreditCard = 'CreditCard',
  CreditCardWebview = 'CreditCardWebview',
  CreditCardList = 'CreditCardList',
  Trading = 'Trading',
  TradingNavigation = 'TradingNavigation',
  TradingBuySellNavigation = 'TradingBuySellNavigation',
  TradingBuySell = 'TradingBuySell',
  WalletDetails = 'WalletDetails',
  DepositAddress = 'DepositAddress',
  WithdrawalAddress = 'WithdrawalAddress',
  Referrals = 'Referrals',
  AccountFee = 'AccountFee',
  IEO = 'IEO',
  IEOinfoDetail = 'IEOinfoDetail',
  IEOinfo = 'IEOinfo',
  IEObuy = 'IEObuy',
  IEOS = 'IEOS',
  IEOSinfo = 'IEOSinfo',
  IEOSbuy = 'IEOSbuy',
  UserTransactionDetail = 'UserTransactionDetail',
  ReferralHistory = 'ReferralHistory',
  Staking = 'Staking',
  StakingScreen = 'StakingScreen',
  StakingHistory = 'StakingHistory',
  StakingAccount = 'StakingAccount',
  StakingUnstake = 'StakingUnstake',
  StakingRedeem = 'StakingRedeem',
  ChooseRegion = 'ChooseRegion',
  RaiseTicket = 'RaiseTicket',
  PDFViewerScreen = 'PDFViewerScreen',
  Newsletter = 'Newsletter',
  GlobianceDeposit = 'GlobianceDeposit',
  GlobianceDepositHistory = 'GlobianceDepositHistory',
  GlobianceDepositWidget = 'GlobianceDepositWidget',
  BusinessBasicInfoView = 'BusinessBasicInfoView',
  BusinessCompanyDocuments = 'BusinessCompanyDocuments',
  LicenseInfoView = 'LicenseInfoView',
  ContactDetailView = 'ContactDetailView',
  ConfirmView = 'ConfirmView',
  StakingFormScreen = 'StakingFormScreen',
  IBansTypeFormScreen = 'IBansTypeFormScreen',
  OrderNActiveScreen = 'OrderNActiveScreen',
  IBansPaymentScreen = 'IBansPaymentScreen',
  CardsLandingScreen = 'CardsLandingScreen',
  StepOneCardsTypeOrderScreen = 'StepOneCardsTypeOrderScreen',
  StepTwoCardsTypeOrderScreen = 'StepTwoCardsTypeOrderScreen',
  StepThreeCardsTypeOrderScreen = 'StepThreeCardsTypeOrderScreen',
  ActiveCardListScreen = 'ActiveCardListScreen',
  CardPaymentScreen = 'CardPaymentScreen',
  CardTopUpScreen = 'CardTopUpScreen',
}